+++ src/main/java/net/minecraft/gametest/framework/GameTestHelper.java
@@ -987,7 +987,8 @@ public class GameTestHelper {
         }
     }
 
-    public int getHeight(final Heightmap.Types heightmap, final int x, final int z) {
+    // MCRe NoiseFarlands: 世界 Y Long 化
+    public long getHeight(final Heightmap.Types heightmap, final long x, final long z) {
         BlockPos absolutePos = this.absolutePos(new BlockPos(x, 0, z));
         return this.relativePos(this.getLevel().getHeightmapPos(heightmap, absolutePos)).getY();
     }
